Refactor: changed some data types

+++ src/main/java/com/example/xbankbackend/dtos/responses/TransactionResponse.java
@@ -4,6 +4,7 @@ import com.example.xbankbackend.enums.CurrencyType;
 import com.example.xbankbackend.enums.TransactionType;
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.time.OffsetDateTime;
 
 @Data
@@ -11,7 +12,7 @@ public class TransactionResponse {
     private TransactionType transactionType;
     private String senderName;
     private String receiverName;
-    private Float amount;
+    private BigDecimal amount;
     private CurrencyType currency;
     private OffsetDateTime transactionDate;
     private String comment;
